i175 Rename packages to follow HSearch conventions
 1. No capital letters in package naming (Java convention)
 2. Implementation/internal packages are named "impl" (HSearch convention)
