+++ core/src/main/java/org/hibernate/search/jsr352/context/jpa/impl/ActiveSessionFactoryRegistryUpdater.java
@@ -1,0 +1,29 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.jsr352.context.jpa.impl;
+
+import org.hibernate.boot.Metadata;
+import org.hibernate.engine.spi.SessionFactoryImplementor;
+import org.hibernate.integrator.spi.Integrator;
+import org.hibernate.service.spi.SessionFactoryServiceRegistry;
+
+/**
+ * @author Yoann Rodiere
+ */
+public class ActiveSessionFactoryRegistryUpdater implements Integrator {
+
+	@Override
+	public void integrate(Metadata metadata, SessionFactoryImplementor sessionFactory, SessionFactoryServiceRegistry serviceRegistry) {
+		ActiveSessionFactoryRegistry.getInstance().register( sessionFactory );
+	}
+
+	@Override
+	public void disintegrate(SessionFactoryImplementor sessionFactory, SessionFactoryServiceRegistry serviceRegistry) {
+		ActiveSessionFactoryRegistry.getInstance().unregister( sessionFactory );
+	}
+
+}
