+++ core/src/main/resources/META-INF/services/org.hibernate.integrator.spi.Integrator
@@ -1,1 +1,1 @@
-org.hibernate.search.jsr352.context.jpa.internal.ActiveSessionFactoryRegistryUpdater
+org.hibernate.search.jsr352.context.jpa.impl.ActiveSessionFactoryRegistryUpdater
